pgwire,jwtauth: remove ConfigureJWTAuth hook indirection
Move the JWTVerifier interface from pgwire into the jwtauth package
where it is implemented. This breaks the artificial import cycle that
previously required a hook-based dependency injection pattern to
connect the two packages.

pgwire and authserver now import jwtauth directly and call
jwtauth.ConfigureJWTAuth as a regular function. This removes:
- The ConfigureJWTAuth hook variable from pgwire
- The noJWTConfigured stub type and its "requires CCL" errors
- The init() in jwtauth that overwrote the hook
- The blank import of jwtauth in ccl_init.go (no longer needed
  since pgwire imports it directly)

Informs #164337

Release note: None
