+++ pkg/server/authserver/BUILD.bazel
@@ -17,6 +17,7 @@ go_library(
         "//pkg/multitenant",
         "//pkg/roachpb",
         "//pkg/security",
+        "//pkg/security/jwtauth",
         "//pkg/security/password",
         "//pkg/security/provisioning",
         "//pkg/security/username",
