+++ pkg/sql/pgwire/BUILD.bazel
@@ -30,6 +30,7 @@ go_library(
         "//pkg/roachpb",
         "//pkg/security",
         "//pkg/security/distinguishedname",
+        "//pkg/security/jwtauth",
         "//pkg/security/password",
         "//pkg/security/provisioning",
         "//pkg/security/sessionrevival",
